scrapers/requirements.txt
requests
beautifulsoup4
pandas
python-dotenv
sec-edgar-downloader
yfinance